NL2SQLWEB/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
python-dotenv

nl2sql @ file:///C:/MyDocs/AppProjects/NL2SQL/dist/nl2sql-0.20.1-py3-none-any.whl

# AutoGen stack
autogen-agentchat==0.4.7
autogen-core==0.4.7
autogen-ext[azure,openai]==0.4.7
aiostream>=0.6.4
grpcio>=1.68.1

# Core dependencies from text_2_sql_core
aiohttp>=3.11.11
aioodbc>=0.5.0
azure-identity>=1.19.0
azure-search>=1.0.0b2
azure-search-documents==11.6.0
jinja2>=3.1.4
networkx>=3.4.2
numpy<2.0.0
openai>=1.55.3
pandas>=2.2.3
pydantic>=2.10.2
python-dotenv>=1.0.1
pyyaml>=6.0.2
rich>=13.9.4
sqlglot[rs]>=25.32.1
tabulate>=0.9.0
tenacity>=9.0.0
typer>=0.14.0

# Additional from autogen
sqlparse>=0.4.4
nltk>=3.8.1
cachetools>=5.5.1
azure-cosmos>=4.9.0

# Database connectors (optional but recommended)
psycopg[binary]>=3.2.3